requirements.txt
# Project dependencies
mcp>=0.9.0
asyncpg>=0.29.0
python-dateutil>=2.8.2
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
